Break apart fetch library into smaller libraries.
